VBox/intnet.h: Corrected alignment padding in INTNETSG.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@104338 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ include/VBox/intnet.h
@@ -348,7 +348,7 @@ typedef struct INTNETSG
     uint16_t volatile   fFlags;
 #if ARCH_BITS == 64
     /** Alignment padding. */
-    uint16_t            uPadding;
+    uint32_t            uPadding;
 #endif
     /** The number of segments allocated. */
     uint16_t            cSegsAlloc;
